add main, and get outputs from main_screen
